feat: NavBar
created the NavBar Component and moved navList from the Header to the NavBar
