+++ src/components/Header/Header.jsx
@@ -1,21 +1,13 @@
 import '../../styles/header.scss'
 import Button from '../Button/Button';
 import { navList } from '../../constants/headerNavList';
-import { useNavigation } from '../../hooks/useNavigation';
+import NavBar from '../NavBar/NavBar';
 
 const Header = () => {
-    const navigate = useNavigation()
+
     return (
         <header className='header'>
-            <nav className='header__nav'>
-                <ul className='nav__ul'>
-                    {navList.map((item) => (
-                        <li key={item.id} onClick={(e) => navigate(e.target.innerText)}>
-                            {item.name}
-                        </li>
-                    ))}
-                </ul>
-            </nav>
+            <NavBar navList={navList} />
             <Button variant={'button-logout'}>Log out</Button>
         </header>
     );
